must specify tag/branch to update

+++ update.sh
@@ -1,10 +1,19 @@
+#!/bin/bash
+
+TAG=$1
+
+if [[ "$TAG" == "" ]]; then
+  echo "no branch or tag specified"
+  exit 1
+fi
+
 cd "$(dirname $0)"
-git clone --depth=1 --branch=master git@github.com:apache/thrift tmp
+git clone --depth=1 --branch=$TAG git@github.com:apache/thrift tmp
 
 # Move old file to an upper level to fit the new requirements of psr-4
 if [ -d "src/Thrift" ]
 then
-	mv src/Thrift/* src/ && rm -Rf src/Thrift
+  mv src/Thrift/* src/ && rm -Rf src/Thrift
 fi
 
 rm -Rf src/Thrift
